classification/requirements.txt
biopython>=1.83
pandas
datasets
tqdm
requests